# Conflicts:
#	android/src/main/kotlin/com/hiennv/flutter_callkit_incoming/CallkitIncomingActivity.kt
#	android/src/main/kotlin/com/hiennv/flutter_callkit_incoming/TransparentActivity.kt
#	example/pubspec.lock
#	ios/Classes/SwiftFlutterCallkitIncomingPlugin.swift

+++ lib/entities/call_kit_params.dart
@@ -5,6 +5,7 @@ import 'ios_params.dart';
 
 part 'call_kit_params.g.dart';
 
+/// Object config for General.
 @JsonSerializable(explicitToJson: true)
 class CallKitParams {
   const CallKitParams({
@@ -30,8 +31,8 @@ class CallKitParams {
   final String? appName;
   final String? avatar;
   final String? handle;
-  final double? type;
-  final double? duration;
+  final int? type;
+  final int? duration;
   final String? textAccept;
   final String? textDecline;
   final String? textMissedCall;
@@ -41,7 +42,8 @@ class CallKitParams {
   final AndroidParams? android;
   final IOSParams? ios;
 
-  factory CallKitParams.fromJson(Map<String, dynamic> json) => _$CallKitParamsFromJson(json);
+  factory CallKitParams.fromJson(Map<String, dynamic> json) =>
+      _$CallKitParamsFromJson(json);
 
   Map<String, dynamic> toJson() => _$CallKitParamsToJson(this);
 }
